pyproject.toml
[project]
name = "semantic-search-module"
version = "0.1.0"
description = "RAG semantic search module for bakery/pastry formulation — queries ingredient technical sheets using cosine similarity over pgvector embeddings."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.30.0",
]
